+++ ui/src/types/v2/dashboards.ts
@@ -1,3 +1,4 @@
+import {HistogramPosition} from 'src/minard'
 import {Color} from 'src/types/colors'
 import {Label} from '@influxdata/influx'
 import {
@@ -122,6 +123,7 @@ export type ViewProperties =
   | MarkdownView
   | EmptyView
   | LogViewerView
+  | HistogramView
 
 export type QueryViewProperties = Extract<
   ViewProperties,
@@ -218,6 +220,18 @@ export interface TableView {
   showNoteWhenEmpty: boolean
 }
 
+export interface HistogramView {
+  type: ViewType.Histogram
+  queries: DashboardQuery[]
+  x: string
+  fill: string
+  position: HistogramPosition
+  binCount: number
+  colors: Color[]
+  note: string
+  showNoteWhenEmpty: boolean
+}
+
 export interface MarkdownView {
   type: ViewType.Markdown
   shape: ViewShape.ChronografV2
@@ -255,6 +269,7 @@ export enum ViewType {
   Table = 'table',
   Markdown = 'markdown',
   LogViewer = 'log-viewer',
+  Histogram = 'histogram',
 }
 
 export interface DashboardFile {